pyproject.toml
[project]
name = "stockpulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.43.6",
    "duckdb>=1.5.2",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
    "yfinance>=1.3.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "UP",  # pyupgrade
    "SIM", # simplify
    "C4",  # comprehensions
    "PIE", # misc improvements
    "N",   # pep8 naming
    "ANN", # type annotations
]

ignore = [
    "ANN101", # self annotation
    "ANN102", # cls annotation
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"

strict = true

warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true

pretty = true
show_error_codes = true
show_column_numbers = true

exclude = ["build/", "dist/", ".venv/"]

[dependency-groups]
dev = ["boto3-stubs[essential]>=1.43.6"]
